changed addon detection
