--- frontend/src/app/actions/browser.js
@@ -1,14 +1,0 @@
-
-// @flow
-
-import type {
-  BrowserState,
-  SetStateAction
-} from "../reducers/browser";
-
-export function setState(payload: BrowserState): SetStateAction {
-  return {
-    type: "SET_STATE",
-    payload: payload
-  };
-}
